Delegate creating Goma instance to GomaFactory
