+++ server/src/domain/goma/Goma.ts
@@ -3,7 +3,6 @@ import SIDE from '../constant/SIDE';
 import { GomaData } from '../../frameworks/data-services/gungi-data';
 import GOMA from '../constant/GOMA';
 import LEVEL from '../constant/LEVEL';
-import Hei from './Hei';
 
 export default abstract class Goma {
   constructor(
@@ -12,17 +11,6 @@ export default abstract class Goma {
     private _name: GOMA,
     private _coordinate: Coordinate,
   ) {}
-
-  static create(level: LEVEL, side: SIDE, name: GOMA, coordinate: Coordinate) {
-    switch (name) {
-      case GOMA.HEI: {
-        return new Hei(level, side, name, coordinate);
-      }
-      default: {
-        throw new Error('Goma name is not exist');
-      }
-    }
-  }
 
   toData(): GomaData {
     return {
