+++ server/src/domain/goma/Hei.ts
@@ -1,3 +1,16 @@
 import Goma from './Goma';
+import SIDE from '../constant/SIDE';
+import Coordinate from '../Coordinate';
+import LEVEL from '../constant/LEVEL';
+import GOMA from '../constant/GOMA';
 
-export default class Hei extends Goma {}
+export default class Hei extends Goma {
+  constructor(
+    _level: LEVEL,
+    _side: SIDE,
+    _name: GOMA,
+    _coordinate: Coordinate,
+  ) {
+    super(_level, _side, _name, _coordinate);
+  }
+}
